Refused apple, drawing button and snake back to original classes.
